module4-task1: + move generate fns to helpers
